Want to be able to specify a new controller (and eventually actions/events too and maybe even customisations) for controls
The idea here is to specify a controller and use that to decide what the specifier is, but this comes with a whole bunch of typing issues to do with the fact that we want the ‘subcontroller’ class to be anonymous.

Need to think of a better way, I think. Keeping this in for now as it possibly supplies the basis for a system that will work.