Make it so that a Session is also the main Controller
This means we don’t set the UI state in the session state any more, but instead return it via the Controller interface.

The various UI bits we have probably should be their own controllers too. Hm, in fact there’s a hidden ‘animation editor’ controller here as we’ll probably want things like file selections and so on.