Update a TODO now I’ve thought about it some more
If a value is calculated on two threads and we return an out-of-date value due to multiple changes, the ‘wrong’ value will never get a proper ‘when changed’ event as things stand as it’ll fire when the newer value changes, leading to problems.

Still want to calculate outside of the lock if possible.