Add a 'layer' to the viewGetCanvasForDrawing function.
There are a couple of ways we can implement layers: we can create them on
the rust side or on the Swift side. The advantage of the swift side is we
can avoid combining them until the layer actually triggers a redraw, and
we can also just store the layers we need to actually render the canvas.